requirements.txt
bokeh
datashader
pandas
python-casacore
numpy